Add support for standard library
This adds a global environment, which is populated once on startup with
the results of evaluating stdlib/stdlib.lisp. This global environment is
then copied over a fresh environment whenever a new one is created. In
effect, this allows writing standard library functions in WELP and make
them available as if they are builtin.

The tests are broken at this point, need to factor out the globals.